敵を複数出現させようとして苦戦中

+++ Enemy.cpp
@@ -19,12 +19,11 @@ void Enemy::Initialize()
 {
 
 	hModel_ = Model::Load("Enemy.fbx");
-
 	Model::SetAnimFrame(hModel_, 1, 100, 1.0f);
 	//モデルハンドル、開始フレーム、終了フレーム、アニメーション速度
 	assert(hModel_ >= 0);
-	float x = (rand() / RAND_MAX) * 5.0f - 10.0f;
-	float z = (rand() / RAND_MAX) * 5.0f - 10.0f;
+	float x = ((float)rand() / RAND_MAX) * 5.0f - 10.0f;
+	float z = ((float)rand() / RAND_MAX) * 5.0f - 10.0f;
 	transform_.position_.x = x;
 	transform_.position_.z = z;
 	SphereCollider* collider = new SphereCollider({ 0,0,0 },0.5f);//コライダーを作る
